// src/modules/auth/auth.schema.ts
import z from "zod";

export const registerSchema = z.object({
    name: z.string().min(1, "Name is required"),
    username: z.string().min(3, "Username must be at least 3 characters"),
    email: z.string().email("Invalid email"),
    password: z.string().min(6, "Password must be at least 6 characters")
})

export const loginSchema = z.object({
    email: z.string().email(),
    password: z.string().min(1, "Password is required")
})

export type RegisterInput = z.infer<typeof registerSchema>;
export type LoginInput = z.infer<typeof loginSchema>